+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## 0.7.5
+### Fixed
+
+ - Breaking: Conflicting description check should now detect all cases thanks to moving it from working on
+   regexes, into working on DFAs.  This change is at the expense of worse error messages.
+
 ## 0.7.4
 ### Fixed